refactor: 파일삭제, 폴더정리
