+++ src/app/layout.tsx
@@ -8,7 +8,7 @@ import Script from "next/script";
 import { ToastProvider } from "./_context/toast-context";
 import Toast from "./_components/noti-toast";
 import { SpeedInsights } from "@vercel/speed-insights/next";
-import ServiceWorkerDebug from "./_components/FCM/ServiceWorkerDebug";
+import ServiceWorkerDebug from "./_components/ServiceWorkerDebug";
 import DebugLogger from "./_components/debug-logger";
 import { LoadingProvider } from "./_context/loading-context";
